Add files via upload
I wrote this article to share my experience as a vegan with as many people as possible. In this article, I put forward my views that a balanced vegan diet is very powerful as a means for preventive medicine and why the diet is inseparable from the Environmental, Social and Governance (ESG) issues that are affecting us. On the benefits to the environment, I describe, based on scientific findings, how this diet is a crucial tool in combating global warming and pollution of our water and land resources.

I strongly believe diseases such as diabetes, high blood pressure, heart-related problems, certain cancers, kidney-related diseases and other diseases are preventable or better managed based on my personal experience and the research findings that I have discussed in this article. In addition to lifestyle changes, adopting renewable energy such as solar and wind power, and other methods, the vegan diet is our partner in humankind's fight against climate change and pollution. 

There is a lot more on vegan diet that we have not understood. I have also listed a number of research topics that graduate students, professors, agencies, government departments, etc. may wish to look into. 

This article does not touch on Buddhism and karma. This shorter version is suitable for those who wish to focus on the science behind the vegan diet. I have a longer version, Plant powered 26 July 2022 for all.pdf, that contains my interpretations of the relationship between the vegan diet and Buddhist beliefs and karma.

This article is currently available free. Please feel free to share with others. Comments are welcome.

The file, Plant powered 26 July 2022 for all.pdf, is also available on GitHub. Please feel free to download and share. 

Thank you, Casey