Add files via upload
I wrote this article to share my experience as a vegan with as many people as possible. In this article, I put forward my views that a balanced vegan diet is very powerful as a means for preventive medicine and why the diet is inseparable from the Environmental, Social and Governance (ESG) issues that are affecting us. On the benefits to the environment, I describe, based on scientific findings, how this diet is a crucial tool in combating global warming and pollution of our water and land resources.

This article also contains my interpretations of the relationship between the vegan diet and Buddhist beliefs and karma. For those who wish to focus on the science behind the vegan diet and do not wish to access content on Buddhism and karma owing to certain religious beliefs, etc., please read the other shorter version, ESG and vegan 26 July 2022 for all.pdf.

I strongly believe diseases such as diabetes, high blood pressure, heart-related problems, certain cancers, kidney-related diseases and other diseases are preventable or better managed based on my personal experience and the research findings that I have discussed in this article. In addition to lifestyle changes, adopting renewable energy such as solar and wind power, and other methods, the vegan diet is our partner in humankind's fight against climate change and pollution. 

This article includes an account of my thoughts when I was a small kid on whether humankind should be eating meat and compassion toward animals. It contains discussions related to karma and why quite a number of Buddhists choose not to eat meat and seafood. 

There is a lot more on vegan diet that we have not understood. I have also listed a number of research topics that graduate students, professors, agencies, government departments, etc. may wish to look into. 

This article is currently available free. Please feel free to share with others. Comments are welcome.

The shorter version, ESG and vegan 26 July 2022 for all, is also available on GitHub. Please feel free to access and share. 

Thank you, Casey